+++ package.json
@@ -4,6 +4,11 @@
     "private": true,
     "homepage": ".",
     "license": "MIT",
+    "scripts": {
+        "start": "wasm-pack build lib && yarn link --cwd lib/pkg && yarn link wave && react-app-rewired start",
+        "build": "wasm-pack build --release lib && yarn link --cwd lib/pkg && yarn link wave && react-app-rewired build",
+        "test": "wasm-pack build lib && yarn link --cwd lib/pkg && yarn link wave && react-app-rewired test --passWithNoTests"
+    },
     "dependencies": {
         "@fortawesome/fontawesome-free": "^5.13.0",
         "@fortawesome/fontawesome-svg-core": "^1.2.28",
@@ -24,45 +29,6 @@
         "typeface-roboto": "^1.1.0",
         "typeface-roboto-mono": "^1.1.0",
         "uuid": "^8.0.0"
-    },
-    "resolutions": {
-        "core-js": "^3.0.0",
-        "source-map-resolve": "^0.6.0",
-        "chokidar": "^3.0.0",
-        "websocket-extensions": "^0.1.4",
-        "execa": "^4.0.0",
-        "serialize-javascript": "^3.1.0",
-        "lodash": "^4.17.16",
-        "ajv": "^6.12",
-        "sockjs": "^0.3"
-    },
-    "scripts": {
-        "start": "wasm-pack build lib && yarn link --cwd lib/pkg && yarn link wave && react-app-rewired start",
-        "build": "wasm-pack build --release lib && yarn link --cwd lib/pkg && yarn link wave && react-app-rewired build",
-        "test": "wasm-pack build lib && yarn link --cwd lib/pkg && yarn link wave && react-app-rewired test --passWithNoTests"
-    },
-    "lint-staged": {
-        "./**/*.{js,jsx,ts,tsx,json}": [
-            "eslint --fix",
-            "prettier --write"
-        ]
-    },
-    "husky": {
-        "hooks": {
-            "pre-commit": "lint-staged"
-        }
-    },
-    "browserslist": {
-        "production": [
-            ">0.2%",
-            "not dead",
-            "not op_mini all"
-        ],
-        "development": [
-            "last 1 chrome version",
-            "last 1 firefox version",
-            "last 1 safari version"
-        ]
     },
     "devDependencies": {
         "@testing-library/jest-dom": "^5.11.0",
@@ -93,5 +59,39 @@
         "styled-components": "^5.1.1",
         "typescript": "^4.2.0",
         "wasm-loader": "^1.3.0"
+    },
+    "resolutions": {
+        "core-js": "^3.0.0",
+        "source-map-resolve": "^0.6.0",
+        "chokidar": "^3.0.0",
+        "websocket-extensions": "^0.1.4",
+        "execa": "^4.0.0",
+        "serialize-javascript": "^3.1.0",
+        "lodash": "^4.17.16",
+        "ajv": "^6.12",
+        "sockjs": "^0.3"
+    },
+    "lint-staged": {
+        "./**/*.{js,jsx,ts,tsx,json}": [
+            "eslint --fix",
+            "prettier --write"
+        ]
+    },
+    "husky": {
+        "hooks": {
+            "pre-commit": "lint-staged"
+        }
+    },
+    "browserslist": {
+        "production": [
+            ">0.2%",
+            "not dead",
+            "not op_mini all"
+        ],
+        "development": [
+            "last 1 chrome version",
+            "last 1 firefox version",
+            "last 1 safari version"
+        ]
     }
 }
