+++ src/types/index.ts
@@ -2,22 +2,10 @@ import { Point, Tile } from "wave";
 
 export { Point, Tile } from "wave";
 
-// export interface Point {
-//     x: number;
-//     y: number;
-// }
-
 export interface Dimension {
     width: number;
     height: number;
 }
-
-// export interface Tile {
-//     location: Point;
-//     value: number;
-//     explored: boolean;
-//     variation: number;
-// }
 
 export enum Direction {
     North,
